microservicios_tareas/requirements.txt
Flask
flask-jwt-extended
pymongo
requests
PyJWT
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-zipkin